#11 Fix: change package root
