+++ src/main/java/com/memesphere/domain/notification/controller/CoinNotificationController.java
@@ -1,11 +1,11 @@
-package com.memesphere.notification.controller;
+package com.memesphere.domain.notification.controller;
 
 
 import com.memesphere.global.apipayload.ApiResponse;
-import com.memesphere.notification.dto.request.NotificationRequest;
-import com.memesphere.notification.dto.response.NotificationListResponse;
-import com.memesphere.notification.dto.response.NotificationResponse;
-import com.memesphere.notification.service.CoinNotificationService;
+import com.memesphere.domain.notification.dto.request.NotificationRequest;
+import com.memesphere.domain.notification.dto.response.NotificationListResponse;
+import com.memesphere.domain.notification.dto.response.NotificationResponse;
+import com.memesphere.domain.notification.service.CoinNotificationService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
